Move: unitDecorator - 폴더 이름 변경 및 함수, 타입 분리
