+++ src/plugins/deleteDecorator/calculate/regex.ts
@@ -1,4 +1,4 @@
-import { numbers } from 'plugins/unitDecorator/utils/unitDecoratorRegex';
+import { numbers } from 'plugins/unitDecorator/calculate/regex';
 
 export const getUnitTotallyMatch = (units = '코'): RegExp =>
   new RegExp(`^[${numbers}]+[${units}]$`, 'gi');
